-set afterkommapicker with last value // default value -changed function and logic of actual weight

+++ app/src/main/java/com/example/liebherr_365_gesundheitsapp/ModulWeight/BmiCalculator.java
@@ -28,7 +28,7 @@ public class BmiCalculator {
         dataSourceData.open();
 
         // getCurrentWeight
-        weight = dataSourceData.getLatestEntry("ModulWeight");
+        weight = (int) dataSourceData.getLatestEntry("ModulWeight");
 
         dataSourceData.close();
